not working
minor changes

rebased

+++ sii/src/byte/byte.c
@@ -1,12 +1,12 @@
 #include "../../includes/byte.h"
 
-int
-getBit(BYTE byte, int whichBit)
+uint8_t
+getBit(BYTE byte, uint8_t whichBit)
 {
     return (byte >> whichBit) & 0x01;
 }
 
 BYTE
-overrideLessSignificantBit(BYTE byte, int bit) {
+overrideLessSignificantBit(BYTE byte, uint8_t bit) {
     return (~1 & byte) | bit;
 }
